+++ tests/image_formats.h
@@ -7,13 +7,13 @@ inline Yttrium::Image make_test_image(bool with_alpha)
 {
 	Yttrium::Buffer buffer(16 * 16 * (with_alpha ? 4 : 3));
 	auto data = &buffer[0];
-	for (size_t y = 0; y < 16; ++y)
+	for (int y = 0; y < 16; ++y)
 	{
-		for (size_t x = 0; x < 16; ++x)
+		for (int x = 0; x < 16; ++x)
 		{
-			uint8_t b = 0;
-			uint8_t g = 0;
-			uint8_t r = 0;
+			int b = 0;
+			int g = 0;
+			int r = 0;
 			if (y < 4)
 			{
 				r = (y * 16 + x) * 4 + 3;
@@ -32,11 +32,11 @@ inline Yttrium::Image make_test_image(bool with_alpha)
 			{
 				b = ((15 - y) * 16 + x) * 4 + 3;
 			}
-			*data++ = b;
-			*data++ = g;
-			*data++ = r;
+			*data++ = static_cast<uint8_t>(b);
+			*data++ = static_cast<uint8_t>(g);
+			*data++ = static_cast<uint8_t>(r);
 			if (with_alpha)
-				*data++ = x * 16 + 15;
+				*data++ = static_cast<uint8_t>(x * 16 + 15);
 		}
 	}
 	return { { 16, 16, with_alpha ? Yttrium::PixelFormat::Bgra : Yttrium::PixelFormat::Bgr, with_alpha ? 32u : 24u }, buffer.data() };
